requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
mcp==1.0.0
hdfs3==0.3.1
requests==2.31.0
aiohttp==3.9.1
jinja2==3.1.2
python-multipart==0.0.6
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
pydantic-settings==2.1.0
PyJWT==2.8.0